bb: Rename BBTagInfo2 parts to BBLV_Info
