+++ bb/src/BBLV_Info.cc
@@ -14,8 +14,8 @@
 #include "bberror.h"
 #include "bbinternal.h"
 #include "bbio.h"
+#include "BBLV_Info.h"
 #include "bbserver_flightlog.h"
-#include "BBTagInfo2.h"
 #include "BBTransferDef.h"
 #include "bbwrkqmgr.h"
 #include "ContribIdFile.h"
