+++ bb/src/BBLV_Info.h
