+++ bb/src/xfer.cc
@@ -37,14 +37,13 @@ namespace bs  = boost::system;
 #include "bbinternal.h"
 #include "bbio_regular.h"
 #include "bbio_BSCFS.h"
+#include "BBLV_Info.h"
 #include "BBLV_Metadata.h"
 #include "bbserver_flightlog.h"
 #include "BBTagID.h"
 #include "BBTransferDef.h"
 #include "bbwrkqmgr.h"
 #include "BBTagID.h"
-#include "BBTagInfo.h"
-#include "BBTagInfo2.h"
 #include "BBTagInfoMap.h"
 #include "ContribIdFile.h"
 #include "ExtentInfo.h"
